+++ integrations/github/CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- towncrier release notes start -->
 
-## 4.0.0-beta (2025-11-18)
+## 4.0.0-beta (2025-11-19)
 
 
 ### Improvements
@@ -18,6 +18,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Introduced helper to centralize repository option construction.
 - During user resync, enrich the personal user with their primary email when needed.
 - Improved HTTP client error messages to include the HTTP method for clearer diagnostics.
+
+
+## 3.3.10-beta (2025-11-19)
+
+
+### Improvements
+
+- dispatch workflow action spec naming improvements
+
+
+## 3.3.9-beta (2025-11-19)
+
+
+### Improvements
+
+- Bumped ocean version to ^0.29.8
 
 
 ## 3.3.8-beta (2025-11-18)